requirements.txt
cryptography==39.0.2
pyOpenSSL==23.0.0
requests==2.28.2
urllib3==1.26.15
